+++ packages/web/src/api/schema.ts
@@ -54,7 +54,7 @@ export interface paths {
           sortDirection?: components['schemas']['SortDirection'];
           page?: number;
           pageSize?: number;
-          includeArchived?: boolean;
+          archived?: components['schemas']['ArchivedFilter'];
         };
         header?: never;
         path?: never;
@@ -979,6 +979,8 @@ export interface components {
       name: null | string;
       version: null | string;
     };
+    /** @enum {unknown} */
+    ArchivedFilter: 'Active' | 'All' | 'Archived' | null;
     ContractRequest: {
       /** Format: uuid */
       customerId: string;